pyproject.toml
[project]
name = "iam57-bot"
version = "1.0.0_beta"
description = "iam57's bot."
readme = "README.md"
requires-python = ">=3.10, <3.12"
dependencies = [
    "nonebot-adapter-onebot>=2.4.3",
    "nonebot2[fastapi]>=2.2.1",
    "nonebot2[websockets]>=2.2.1",
    "httpx>=0.27.0",
    "aiosqlite>=0.17.0",
    "tortoise-orm[aiosqlite]>=0.21.5",
    "nonebot-plugin-alconna>=0.45.0",
]

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugins = []
plugin_dirs = ["iam57_bot/plugins"]
builtin_plugins = []

[tool.pdm]
